+++ packages/core/src/index.ts
@@ -1,8 +1,9 @@
 import Form, { FormProps, FormState, IChangeEvent } from './components/Form';
+import RichDescription, { RichDescriptionProps } from './components/RichDescription';
 import withTheme, { ThemeProps } from './withTheme';
 import getDefaultRegistry from './getDefaultRegistry';
 
-export type { FormProps, FormState, IChangeEvent, ThemeProps };
+export type { FormProps, FormState, IChangeEvent, ThemeProps, RichDescriptionProps };
 
-export { withTheme, getDefaultRegistry };
+export { withTheme, getDefaultRegistry, RichDescription };
 export default Form;