+++ packages/core/test/ArraySnap.test.tsx
@@ -1,4 +1,5 @@
-import Form from '../src';
 import { arrayTests } from '@rjsf/snapshot-tests';
+
+import Form from '../src';
 
 arrayTests(Form);