+++ packages/react-bootstrap/src/CheckboxWidget/CheckboxWidget.tsx
@@ -49,10 +49,7 @@ export default function CheckboxWidget<
 
   const description = options.description || schema.description;
   return (
-    <Form.Group
-      className={`checkbox ${disabled || readonly ? 'disabled' : ''}`}
-      aria-describedby={ariaDescribedByIds<T>(id)}
-    >
+    <Form.Group className={disabled || readonly ? 'disabled' : ''} aria-describedby={ariaDescribedByIds<T>(id)}>
       {!hideLabel && !!description && (
         <DescriptionFieldTemplate
           id={descriptionId<T>(id)}